backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
PyMuPDF==1.24.10
groq==0.11.0
pydantic==2.9.2
python-dotenv==1.0.1
